Move the reactive list classes (WatchedArray, WatchedSet, etc.) into @dynein/state and move hyperfor (renamed to D.addFor) into @dynein/dom. This will simplify the installation and usage of Dynein, because both the watched-builtins classes and fast list rendering will generally be needed in real-world Dynein apps.
The @dynein/watched-builtins and @dynein/hyperfor packages will now be deleted/deprecated.

The original reason I made them separate packages was to decrease the size of the Dynein core. But as noted above, they will probably be needed in real development, and so excluding them from the advertised base Dynein bundle size would be somewhat misleading.

Modern web apps are generally bundled with a tree-shaking bundler, and so if these APIs are not used, they won't increase the bundle size anyway..